Revert "feat(calendar): separate calendarMonthOnlyView"
This reverts commit 3fe44d771b58add7b1e0b40b627e9031816b8ebc.
