+++ packages/calendar/src/components/calendar-mobile/Component.tsx
@@ -1,17 +1,39 @@
-import React, { forwardRef, useState } from 'react';
+import React, { forwardRef, useMemo, useRef, useState } from 'react';
 import mergeRefs from 'react-merge-refs';
+import { Virtuoso } from 'react-virtuoso';
 import { ResizeObserver as ResizeObserverPolyfill } from '@juggle/resize-observer';
 import cn from 'classnames';
+import endOfDay from 'date-fns/endOfDay';
+import isAfter from 'date-fns/isAfter';
+import isSameDay from 'date-fns/isSameDay';
+import isSameMonth from 'date-fns/isSameMonth';
+import isThisMonth from 'date-fns/isThisMonth';
+import lastDayOfMonth from 'date-fns/lastDayOfMonth';
+import startOfDay from 'date-fns/startOfDay';
+import startOfMonth from 'date-fns/startOfMonth';
 
 import { ButtonMobile } from '@alfalab/core-components-button/mobile';
 import { ModalMobile } from '@alfalab/core-components-modal/mobile';
 import { getDataTestId } from '@alfalab/core-components-shared';
+import { Typography } from '@alfalab/core-components-typography';
 
 import { CalendarDesktop } from '../../desktop';
-import { isRangeValue, WEEKDAYS } from '../../utils';
+import { Month } from '../../typings';
+import { useCalendar } from '../../useCalendar';
+import { useRange } from '../../useRange';
+import {
+    addonArrayToHashTable,
+    dateArrayToHashTable,
+    generateMonths,
+    generateWeeks,
+    isRangeValue,
+    limitDate,
+    monthName,
+    WEEKDAYS,
+} from '../../utils';
+import { DaysTable } from '../days-table';
 
-import { CalendarMonthOnlyView } from './components/calendar-month-only-view';
-import { CalendarMobileProps } from './typings';
+import { CalendarContentProps, CalendarMobileProps } from './typings';
 
 import backdropTransitionStyles from './backdrop-transitions.module.css';
 import styles from './index.module.css';
@@ -21,6 +43,261 @@ import transitionStyles from './transitions.module.css';
 if (typeof window !== 'undefined' && !window.ResizeObserver) {
     window.ResizeObserver = ResizeObserverPolyfill;
 }
+
+export const CalendarMonthOnlyView = ({
+    value,
+    mode = 'single',
+    rangeBehavior = 'clarification',
+    month: monthTimestamp,
+    minDate: minDateTimestamp,
+    maxDate: maxDateTimestamp,
+    defaultMonth: defaultMonthTimestamp,
+    offDays,
+    events,
+    holidays,
+    onChange,
+    onMonthTitleClick,
+    selectedFrom,
+    selectedTo,
+    yearsAmount = 3,
+    dayAddons,
+    shape = 'rounded',
+    scrollableContainer,
+    clickableMonth,
+}: CalendarContentProps & {
+    /**
+     * FIXME нужно сделать для компонента CalendarMonthOnlyView отдельный тип пропсов, т.к. тип CalendarContentProps intersection для типа CalendarMobileProps
+     * FIXME это приводит к тому, что в доку сторибука попадают типы пропсов, которые нужны для работы компонента CalendarMonthOnlyView, но не нужны для компонента CalendarMobile
+     * TODO Вынести компонент CalendarMonthOnlyView в отдельный файл
+     */
+    clickableMonth?: boolean;
+}) => {
+    const range = useRange({
+        mode,
+        value,
+        selectedFrom,
+        selectedTo,
+        rangeBehavior,
+        onChange,
+    });
+
+    const month = useMemo(
+        () => (monthTimestamp ? new Date(monthTimestamp) : undefined),
+        [monthTimestamp],
+    );
+
+    const minDate = useMemo(
+        () => (minDateTimestamp ? startOfDay(minDateTimestamp) : undefined),
+        [minDateTimestamp],
+    );
+
+    const maxDate = useMemo(() => {
+        // блокируем последующие дни после текущего
+        if (clickableMonth && !maxDateTimestamp) {
+            return new Date();
+        }
+
+        return maxDateTimestamp ? endOfDay(maxDateTimestamp) : undefined;
+    }, [maxDateTimestamp, clickableMonth]);
+
+    const selected = useMemo(
+        () => (range.value ? new Date(range.value) : undefined),
+        [range.value],
+    );
+
+    const startingDate = useRef(range.value);
+
+    const defaultMonth = useMemo(
+        () =>
+            startOfMonth(
+                selected ||
+                    limitDate(
+                        defaultMonthTimestamp || Date.now(),
+                        minDateTimestamp,
+                        maxDateTimestamp,
+                    ),
+            ),
+        [defaultMonthTimestamp, maxDateTimestamp, minDateTimestamp, selected],
+    );
+
+    const { activeMonth, highlighted, getDayProps } = useCalendar({
+        month,
+        defaultMonth,
+        view: 'months',
+        minDate,
+        maxDate,
+        selected,
+        offDays,
+        events,
+        onChange: range.onChange,
+        dayAddons,
+    });
+
+    const activeMonths = useMemo(() => {
+        const eventsMap = dateArrayToHashTable(events || []);
+        const offDaysMap = dateArrayToHashTable(offDays || []);
+        const holidaysMap = dateArrayToHashTable(holidays || []);
+        const dayAddonsMap = addonArrayToHashTable(dayAddons || []);
+
+        const prevMonths: Month[] = [];
+        const nextMonths: Month[] = [];
+
+        const date = startingDate.current ? new Date(startingDate.current) : new Date();
+        const currentYear = date.getFullYear();
+        const currYearMonths = generateMonths(date, {});
+
+        for (let i = 0; i < yearsAmount; i++) {
+            const prevYear = date.setFullYear(currentYear - (i + 1));
+            const nextYear = date.setFullYear(currentYear + (i + 1));
+
+            const prevYearMonths = generateMonths(new Date(prevYear), {});
+            const nextYearMonths = generateMonths(new Date(nextYear), {});
+
+            prevMonths.unshift(...prevYearMonths);
+            nextMonths.push(...nextYearMonths);
+        }
+
+        const generatedMonths = [...prevMonths, ...currYearMonths, ...nextMonths];
+
+        return generatedMonths.map((item) => ({
+            ...item,
+            weeks: generateWeeks(item.date, {
+                minDate,
+                maxDate,
+                selected,
+                eventsMap,
+                offDaysMap,
+                holidaysMap,
+                dayAddonsMap,
+            }),
+            title: `${monthName(item.date)} ${item.date.getFullYear()}`,
+        }));
+    }, [events, offDays, holidays, dayAddons, minDate, maxDate, yearsAmount, selected]);
+
+    const initialMonthIndex = useMemo(() => {
+        const date = range.value || range.selectedFrom || activeMonth.getTime() || Date.now();
+
+        return activeMonths.findIndex((m) => isSameMonth(date, m.date));
+    }, [range.value, range.selectedFrom, activeMonth, activeMonths]);
+
+    // заголовок должен становиться активным если выбран весь доступный период в месяце
+    const isMonthActive = (currentMonthIndex: number) => {
+        if (value && isRangeValue(value)) {
+            const { date: initialMonthDate } = activeMonths[initialMonthIndex];
+            const { date: currentMonthDate } = activeMonths[currentMonthIndex];
+
+            const firstAvailableDayOfMonth = startOfMonth(initialMonthDate).getTime();
+            /**
+             * последний доступный день месяца в timestamp
+             * представлен в виде последнего календарного дня, либо в виде текущей даты для актуального месяца
+             */
+            const lastAvailableDayOfMonth = isThisMonth(initialMonthDate)
+                ? startOfDay(new Date()).getTime()
+                : lastDayOfMonth(initialMonthDate).getTime();
+            const { dateFrom, dateTo } = value;
+
+            if (
+                dateFrom &&
+                dateTo &&
+                isSameMonth(initialMonthDate, currentMonthDate) &&
+                isSameDay(firstAvailableDayOfMonth, dateFrom) &&
+                isSameDay(lastAvailableDayOfMonth, dateTo)
+            ) {
+                return true;
+            }
+        }
+
+        return false;
+    };
+
+    const handleClickMonthLabel = (index: number) => {
+        if (onChange) {
+            const { date } = activeMonths[index];
+            const firstAvailableDayOfMonth = startOfMonth(date).getTime();
+            const lastAvailableDayOfMonth = isThisMonth(date)
+                ? startOfDay(new Date()).getTime()
+                : lastDayOfMonth(date).getTime();
+
+            if (isMonthActive(index)) {
+                onChange();
+            } else {
+                onChange(firstAvailableDayOfMonth, lastAvailableDayOfMonth);
+            }
+        }
+    };
+
+    const getMonthLabel = (index: number, isClickableMonth?: boolean) => {
+        if (isClickableMonth) {
+            return (
+                <Typography.Text className={styles.monthTitle} view='primary-small' color='primary'>
+                    {activeMonths[index].title}
+                </Typography.Text>
+            );
+        }
+
+        return `\u00A0${activeMonths[index].title}\u00A0`;
+    };
+
+    const renderMonth = (index: number) => {
+        const isAfterDate = isAfter(activeMonths[index].date, activeMonth);
+
+        return (
+            <div className={styles.daysTable} id={`month-${index}`}>
+                {onMonthTitleClick ? (
+                    /* eslint-disable-next-line jsx-a11y/click-events-have-key-events */
+                    <span
+                        className={styles.month}
+                        onClick={onMonthTitleClick}
+                        tabIndex={0}
+                        role='button'
+                    >
+                        {activeMonths[index].title}
+                    </span>
+                ) : (
+                    // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions
+                    <span
+                        className={cn(styles.month, {
+                            ...(clickableMonth && {
+                                [styles.clickableMonth]: true,
+                                [styles.rectangular]: shape === 'rectangular',
+                                [styles.active]: isMonthActive(index),
+                                [styles.disabled]: isAfterDate,
+                            }),
+                        })}
+                        {...(clickableMonth && { onClick: () => handleClickMonthLabel(index) })}
+                    >
+                        {getMonthLabel(index, clickableMonth)}
+                    </span>
+                )}
+                <DaysTable
+                    withTransition={false}
+                    weeks={activeMonths[index].weeks}
+                    activeMonth={activeMonth}
+                    selectedFrom={range.selectedFrom}
+                    selectedTo={range.selectedTo}
+                    getDayProps={getDayProps}
+                    highlighted={highlighted}
+                    hasHeader={false}
+                    responsive={true}
+                    shape={shape}
+                />
+            </div>
+        );
+    };
+
+    return (
+        <Virtuoso
+            totalCount={activeMonths.length}
+            itemContent={renderMonth}
+            initialTopMostItemIndex={{ index: initialMonthIndex ?? 0, align: 'center' }}
+            increaseViewportBy={500}
+            itemSize={(el) => el.getBoundingClientRect().height + 32}
+            customScrollParent={scrollableContainer}
+            useWindowScroll={true}
+            className={styles.virtuoso}
+        />
+    );
+};
 
 export const CalendarMonthOnlyViewHeader = () => (
     <table className={styles.dayNames}>
