--- packages/calendar/src/components/calendar-mobile/components/calendar-month-only-view/index.ts
@@ -1,1 +1,0 @@
-export { CalendarMonthOnlyView } from './calendarMonthOnlyView';
