+++ packages/calendar/src/components/calendar-mobile/index.ts
@@ -1,3 +1,2 @@
 export * from './Component';
 export * from './typings';
-export * from './components/calendar-month-only-view';
